0.5.3
* Move all math utilities to math module
* Add implementation for a linear algebra library
* Use 64 bit integers
  * Fix integer overflow in renderer progress output
* Improve camera builder interface
* Fix uv coordinates for infinite planes
* Add simple Mandelbrot texture
* Use libcurl instead of ureq
* Remove dependencies on uuid, tempfile, and regex crates

+++ src/efloat.rs
@@ -2,7 +2,7 @@ use std::ops::{Add, Div, Mul, Neg, Sub};
 use crate::Float;
 
 /// Used for error estimation in manually propagated floating point errors
-pub fn gamma(n: i32) -> Float {
+pub fn gamma(n: u64) -> Float {
     let n = n as Float;
     (n * Float::EPSILON) / (1.0 - n * Float::EPSILON)
 }
