+++ src/renderer/task.rs
@@ -1,15 +1,16 @@
 use super::*;
 
-pub type RenderTaskExecutor = dyn Fn(RenderTask, &RefCell<Xorshift>, &Camera, &Scene)
-                                     -> RenderTaskResult + Send + Sync + 'static;
+pub type RenderTaskExecutor =
+    dyn Fn(RenderTask, &RefCell<Xorshift>, &Camera, &Scene)
+           -> RenderTaskResult + Send + Sync + 'static;
 
 pub struct RenderTask {
     pub tile: Option<FilmTile>,
-    pub samples: u32,
+    pub samples: u64,
 }
 
 impl RenderTask {
-    pub fn new(tile: FilmTile, samples: u32) -> Self {
+    pub fn new(tile: FilmTile, samples: u64) -> Self {
         Self { tile: Some( tile ), samples }
     }
 
@@ -20,11 +21,11 @@ impl RenderTask {
 
 pub struct RenderTaskResult {
     pub tile: Option<FilmTile>,
-    pub num_rays: i32,
+    pub num_rays: u64,
 }
 
 impl RenderTaskResult {
-    pub fn new(tile: FilmTile, num_rays: i32) -> Self {
+    pub fn new(tile: FilmTile, num_rays: u64) -> Self {
         Self { tile: Some(tile), num_rays, }
     }
